+++ src/GameCollector.PkgHandlers.Winget/WindowsPackageManager/ClsidContext.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace GameCollector.PkgHandlers.Winget.WindowsPackageManager;
+
+public enum ClsidContext
+{
+    // Production CLSID Guids
+    Prod,
+
+    // Development CLSID Guids
+    Dev,
+}